# config/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Watson wherespace",
        "description": "",
        "version": "1.0.0",
        "contact":{
        "name": "OpenNTF",
        "url": "http://www.openntf.org",
        "email": "info@openntf.org"
        }
    },
    "host": "api.rabobank.n",
    "schemes": [
        "https"
    ],
    "basePath": "/v2",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/events/{eventId}": {
            "get": {
                "summary": "Get the event by his id",
                "description": "Return the Event by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "The ID of the event.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Event"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "400": {
                        "description": "Invalid event ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create the event by his id",
                "description": "Create the Event by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "The ID of the event.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Event"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "400": {
                        "description": "Invalid event ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the event by his id",
                "description": "Update the Event by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "The ID of the event.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Event"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "400": {
                        "description": "Invalid event ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete the event by his id",
                "description": "Delete the Event by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "The ID of the event.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Event"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "400": {
                        "description": "Invalid event ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/events/person":{
            "get":{
                "summary": "Get events by person",
                "description": "Get the event by personId",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "email",
                        "description": "Fetched all the events by person email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Event"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/events/person/eventdate":{
            "get":{
                "summary": "Get events by person",
                "description": "Get the events by email and date",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "email",
                        "description": "Fetched all the events by person email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "date",
                        "description": "Fetched all the events by certain date",
                        "required": true,
                        "type": "number"
                    }
                ],
                "tags": [
                    "Event"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email and date supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Event": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "personId": {
                    "type": "string"
                },
                "locationId": {
                    "type": "string"
                },
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "latitude":{
                    "type": "number",
                    "format": "float"
                },
                "longitude":{
                    "type": "number",
                    "format": "float"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        },
        "Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}